+++ PhotonBeam/third-party-helper/tiny-gltf-helper/GltfScene.cpp
